configs should now work i hope. same with compa items drops

+++ src/main/java/com/robertx22/loot/gens/AwakenRuneWordLootGen.java
@@ -23,7 +23,7 @@ public class AwakenRuneWordLootGen extends BaseLootGen {
 
     @Override
     public float BaseChance() {
-        return ModConfig.INSTANCE.DropRates.AWAKEN_RUNEWORD_DROPRATE.get();
+        return ModConfig.INSTANCE.DropRates.AWAKEN_RUNEWORD_DROPRATE.get().floatValue();
 
     }
