+++ src/main/java/com/robertx22/loot/gens/RuneLootGen.java
@@ -30,7 +30,7 @@ public class RuneLootGen extends BaseLootGen {
 
     @Override
     public float BaseChance() {
-        return ModConfig.INSTANCE.DropRates.RUNE_DROPRATE.get();
+        return ModConfig.INSTANCE.DropRates.RUNE_DROPRATE.get().floatValue();
     }
 
     @Override
